lib/rst/interp_float clean-up:  Use off_t for offsets  Fully type function pointers  Fix type of check_points field in struct interp_params to match   IL_check_at_points_2d()
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@38715 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/rst/interp_float/segmen2d.c
@@ -21,8 +21,8 @@ int IL_interp_segments_2d(struct interp_params *params, struct tree_info *info,	
 			  double *gmin, double *gmax,	/* min and max inperp. slope val. */
 			  double *c1min, double *c1max, double *c2min, double *c2max,	/* min and max interp. curv. val. */
 			  double *ertot,	/* total interplating func. error */
-			  int totsegm,	/* total number of segments */
-			  int offset1,	/* offset for temp file writing */
+			  int totsegm,		/* total number of segments */
+			  off_t offset1,	/* offset for temp file writing */
 			  double dnorm)
 /*
    Recursively processes each segment in a tree by
